scrapers/requirements.txt
groq
requests
urllib3
beautifulsoup4
flask
flask-cors
nltk
python-dotenv
